+++ packages/cli/src/commands/init.ts
@@ -1,5 +1,42 @@
 import * as p from '@clack/prompts'
-import { CONFIG_FILE_NAME, Config, writeConfig } from '../config/config'
+import {
+  CONFIG_FILE_NAME,
+  Config,
+  getCssFramework,
+  getImportAliases,
+  getJsFramework,
+  writeConfig,
+} from '../config/config'
+import { saveComponentToFile } from '../helpers/save-file'
+
+const downloadHelpers = async (options: {
+  cssFramework: Config['cssFramework']
+  jsFramework: Config['jsFramework']
+}): Promise<{ filename: string; content: string }[]> => {
+  const { cssFramework, jsFramework } = options
+  const helpersUrl = `https://park-ui.com/registry/${cssFramework}/${jsFramework}/helpers/index.json`
+  const components = await fetch(helpersUrl)
+    .then((res) => res.json())
+    .then((res) => res.files)
+    .catch((e) => {
+      throw new Error(`Failed to download ${jsFramework} ${cssFramework} helpers\n${e?.message}`)
+    })
+  return components
+}
+
+export const addHelpers = async () => {
+  const cssFramework = await getCssFramework()
+  const jsFramework = await getJsFramework()
+  const { utilsImportAlias } = await getImportAliases()
+
+  const components = await downloadHelpers({
+    cssFramework,
+    jsFramework,
+  })
+  components.forEach(({ filename, content }) => {
+    saveComponentToFile(utilsImportAlias, filename, content)
+  })
+}
 
 const getInitialConfig = async (): Promise<Config> => {
   const config = await p.group(
@@ -26,7 +63,7 @@ const getInitialConfig = async (): Promise<Config> => {
       importAliasComponents: () =>
         p.text({
           message: 'What is your import alias for components?',
-          initialValue: '~/components',
+          initialValue: '~/lib/components',
         }),
       importAliasUtils: () =>
         p.text({ message: 'What is your import alias for utils?', initialValue: '~/lib/utils' }),
@@ -36,7 +73,7 @@ const getInitialConfig = async (): Promise<Config> => {
         return p.confirm({ message: 'Do you want to use server components?' })
       },
       confirm: () => p.confirm({ message: `Write the config to ${CONFIG_FILE_NAME} ?` }),
-      outro: () =>
+      outro: async () =>
         p.note(
           "🚀 You're all set now. Happy hacking! \n \nYou can start now adding your first component.\nFor example, run `park-ui add button`.",
         ),
@@ -48,6 +85,11 @@ const getInitialConfig = async (): Promise<Config> => {
       },
     },
   )
+
+  // TODO show spinner while downloading the files before ending the group https://github.com/natemoo-re/clack/issues/152
+  // maybe consider another lib than clack??
+
+  await addHelpers()
 
   return {
     cssFramework: config.cssFramework as any,